Move SourceInvite model from images app to sources app.
